refactor(profile): move navLinks to constants and getLinkClass to utils
