+++ src/pages/Profile/constants.ts
@@ -1,0 +1,5 @@
+export const navLinks = [
+    { path: "contact", label: "Contact Information" },
+    { path: "security", label: "Account Security" },
+    { path: "orders", label: "Orders" },
+];
